Converts map and filters to sibling components
