+++ components/Map.js
@@ -74,7 +74,7 @@ class Map extends React.Component {
     });
 
     // Set features to load based on the default from and to dates
-    const { allModesSelected } = this.makeFeaturesQuery(
+    const { allModesSelected } = this.props.makeFeaturesQuery(
       this.props.modeSelection,
       this.props.fromDate,
       this.props.toDate
@@ -110,8 +110,9 @@ class Map extends React.Component {
       });
   };
 
+  // Update query and features when new selections are made
   componentWillReceiveProps({ modeSelection, fromDate, toDate }) {
-    const { allModesSelected, oneModeSelected } = this.makeFeaturesQuery(
+    const { allModesSelected, oneModeSelected } = this.props.makeFeaturesQuery(
       modeSelection,
       fromDate,
       toDate
@@ -130,6 +131,15 @@ class Map extends React.Component {
     }
   }
 
+  // Update features when user makes new selections
+  updateFeatures = query => {
+    this.featureLayer.setWhere(query, () => {
+      // use _currentSnapshot to update the feature count after re-setting the map
+      const numFeatures = this.featureLayer._currentSnapshot.length;
+      this.setState({ crashCounts: numFeatures });
+    });
+  };
+
   // Set popup for features
   bindPopUp = (feature, layer) => {
     // Format dispatch timestamp to be readable
@@ -142,44 +152,18 @@ class Map extends React.Component {
     const popupContent = `<div style="font-family:'Roboto'"><p><strong>Crash Type:</strong> ${
       feature.properties.mode_type
     }<br><strong>Dispatch Time Stamp:</strong> ${formattedDate}
-      <br><strong>Location Type:</strong> ${
-        feature.properties.location_type
-      }</p></div>`;
+          <br><strong>Location Type:</strong> ${
+            feature.properties.location_type
+          }</p></div>`;
 
     // Add our popups to the features
     layer.bindPopup(popupContent);
   };
 
-  makeFeaturesQuery = (modeSelection, fromDate, toDate) => {
-    // set query for when "all" modes are selected
-    const allModesSelected = `dispatch_ts >=
-      '${fromDate}' AND 
-      dispatch_ts <= '${toDate}'`;
-
-    // set query for when one mode is selected
-    const oneModeSelected = `mode_type =
-      '${modeSelection}' AND 
-      dispatch_ts >= '${fromDate}'AND 
-      dispatch_ts <= '${toDate}'`;
-
-    return {
-      allModesSelected,
-      oneModeSelected,
-    };
-  };
-
-  updateFeatures = query => {
-    this.featureLayer.setWhere(query, () => {
-      // use _currentSnapshot to update the feature count after re-setting the map
-      const numFeatures = this.featureLayer._currentSnapshot.length;
-      this.setState({ crashCounts: numFeatures });
-    });
-  };
-
   render() {
     return (
       <div>
-        <div style={{ height: '600px' }} ref={this.setMapEl}>
+        <div style={{ height: '100vh' }} ref={this.setMapEl}>
           <div
             style={{ zIndex: 1000, position: 'relative', fontFamily: 'Roboto' }}
           >
@@ -194,10 +178,11 @@ class Map extends React.Component {
   }
 }
 
+export default Map;
+
 Map.propTypes = {
-  modeSelection: PropTypes.string,
   fromDate: PropTypes.string,
   toDate: PropTypes.string,
+  modeSelection: PropTypes.string,
+  makeFeaturesQuery: PropTypes.func,
 };
-
-export default Map;
